Fix Profile medium rendering
