+++ src/profile/profile.unit.tsx
@@ -1,20 +1,21 @@
 import React from 'react'
-import { shallow } from 'enzyme'
+import { mount } from 'enzyme'
 
 import Profile from 'profile'
 import Item from '_utils/item'
 import Avatar from 'avatar'
 import Rating from 'rating'
-import Text, { TextDisplayType } from 'text'
+import { TextDisplayType } from 'text'
+import TextBody from 'typography/body'
 
 const defaultProps = {
   title: 'Jack Sparrow',
 }
 describe('Profile default', () => {
-  let defaultProfile
+  let defaultProfile: any
 
   beforeEach(() => {
-    defaultProfile = shallow(<Profile {...defaultProps} />)
+    defaultProfile = mount(<Profile {...defaultProps} />)
   })
 
   it('Should pass a title prop to Item', () => {
@@ -28,7 +29,7 @@ describe('Profile default', () => {
   it('Should pass a title prop to Item', () => {
     expect(defaultProfile.find(Item).prop('leftTitleDisplay')).toEqual(TextDisplayType.TITLE)
 
-    const profileMedium = shallow(<Profile {...defaultProps} isMedium />)
+    const profileMedium = mount(<Profile {...defaultProps} isMedium />)
     expect(profileMedium.find(Item).prop('leftTitleDisplay')).toEqual(TextDisplayType.DISPLAY1)
   })
 
@@ -39,48 +40,61 @@ describe('Profile default', () => {
 
 describe('Profile', () => {
   it('Should have `ariaLabel`', () => {
-    const profile = shallow(<Profile {...defaultProps} ariaLabel="testLabel" />)
+    const profile = mount(<Profile {...defaultProps} ariaLabel="testLabel" />)
     expect(profile.find(Item).prop('ariaLabel')).toEqual('testLabel')
   })
 
   it('Should display info if no rating is provided', () => {
-    const profile = shallow(<Profile {...defaultProps} info="fhtagn" />)
-    expect(profile.find(Item).prop('leftBody')).toEqual(<Text>fhtagn</Text>)
+    const profile = mount(<Profile {...defaultProps} info="test label" />)
+    expect(
+      profile
+        .find(Item)
+        .find(TextBody)
+        .text(),
+    ).toEqual('test label')
   })
 
   it('Should display the rating over info if both are provided', () => {
-    const profile = shallow(<Profile {...defaultProps} ratings="2" info="fhtagn" />)
-    expect(profile.find(Item).prop('leftBody')).toEqual(<Rating ratings="2" />)
+    const profile = mount(<Profile {...defaultProps} ratings={2} info="test label" />)
+    const rating = profile.find(Item).find(Rating)
+    expect(rating.exists()).toBeTruthy()
+    expect(rating.exists()).toBeTruthy()
+    expect(rating.prop('ratings')).toBe(2)
   })
 
   it('Should pass a rightAddon prop to Item', () => {
     const pictureUrl = 'https://s3.amazonaws.com/37assets/svn/1065-IMG_2529.jpg'
-    const profile = shallow(<Profile {...defaultProps} picture={pictureUrl} />)
-    expect(profile.find(Item).prop('rightAddon')).toEqual(<Avatar image={pictureUrl} />)
+    const profile = mount(<Profile {...defaultProps} picture={pictureUrl} />)
+    const avatar = profile.find(Item).find(Avatar)
+    expect(avatar.exists()).toBeTruthy()
+    expect(avatar.prop('image')).toBe(pictureUrl)
   })
 
   it('Should pass all picture props to Avatar', () => {
     const pictureUrl = 'https://s3.amazonaws.com/37assets/svn/1065-IMG_2529.jpg'
-    const profile = shallow(
-      <Profile {...defaultProps} picture={pictureUrl} alt="fhtagn" isMedium isIdChecked />,
+    const profile = mount(
+      <Profile {...defaultProps} picture={pictureUrl} alt="alt_text" isMedium isIdChecked />,
     )
-    expect(profile.find(Item).prop('rightAddon')).toEqual(
-      <Avatar image={pictureUrl} alt="fhtagn" isMedium isIdChecked />,
-    )
+    const avatar = profile.find(Item).find(Avatar)
+    expect(avatar.exists()).toBeTruthy()
+    expect(avatar.prop('image')).toBe(pictureUrl)
+    expect(avatar.prop('alt')).toBe('alt_text')
+    expect(avatar.prop('isMedium')).toBe(true)
+    expect(avatar.prop('isIdChecked')).toBe(true)
   })
 
   it('Should pass a chevron prop to Item when isLink is true', () => {
-    const profile = shallow(<Profile {...defaultProps} isLink />)
+    const profile = mount(<Profile {...defaultProps} isLink />)
     expect(profile.find(Item).prop('chevron')).toBe(true)
   })
 
   it('Should pass a chevron prop to Item when href is present', () => {
-    const profile = shallow(<Profile {...defaultProps} href="#" />)
+    const profile = mount(<Profile {...defaultProps} href="#" />)
     expect(profile.find(Item).prop('chevron')).toBe(true)
   })
 
   it('Should pass a chevron prop to Item when href is present', () => {
-    const profile = shallow(<Profile {...defaultProps} onClick={() => null} />)
+    const profile = mount(<Profile {...defaultProps} onClick={() => null} />)
     expect(profile.find(Item).prop('chevron')).toBe(true)
   })
 })
